# internal/openapi/spec.yaml
openapi: 3.1.0
info:
  title: System Wrangler API
  version: v1
  description: |
    HTTP API exposed by the System Wrangler backend. Every protected
    endpoint requires the `sw_session` cookie issued by
    `POST /api/auth/login` (or `POST /api/auth/setup` on a fresh
    install). State-changing requests must also carry the
    `X-Sw-Csrf: 1` header — the CSRF middleware rejects authenticated
    POST/PUT/PATCH/DELETE without it.

    This spec is the source of truth for the API surface. A drift test
    (`cmd/server/openapi_drift_test.go`) compares it to the routes
    registered on the production mux so the two cannot rot apart.
servers:
  - url: /
    description: Same-origin (the SPA and the API share the host).
...
security:
  - sessionCookie: []
paths:
...
  /api/admin/users:
    get:
      tags: [users]
      summary: List every user account.
      responses:
        "200":
          description: All users.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [users]
      summary: Create a user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: User created. The user starts with zero role assignments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: Username already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/users/{id}:
    patch:
      tags: [users]
      summary: Update an account (currently `disabled` only).
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                disabled: { type: boolean }
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Delete an account.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: User deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/users/{id}/password:
    post:
      tags: [users]
      summary: Admin password reset (sets `mustChangePassword` on the target).
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        "204": { description: Password reset. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/users/{id}/totp/reset:
    post:
      tags: [users]
      summary: Admin-reset TOTP for a user (disables 2FA and wipes trusted devices).
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: 2FA disabled on the target account. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/audit:
    get:
      tags: [audit]
      summary: Page through audit-log rows.
      parameters:
        - { in: query, name: actor_id, schema: { type: string } }
        - { in: query, name: actor_label, schema: { type: string }, description: "Case-insensitive substring match." }
        - { in: query, name: action, schema: { type: string }, description: "Exact match, or prefix when trailing '*'." }
        - { in: query, name: target_kind, schema: { type: string } }
        - { in: query, name: target_id, schema: { type: string } }
        - { in: query, name: target_label, schema: { type: string }, description: "Case-insensitive substring match." }
        - { in: query, name: outcome, schema: { type: string, enum: [success, failure, denied] } }
        - { in: query, name: request_id, schema: { type: string } }
        - { in: query, name: since, schema: { type: integer }, description: "Integer milliseconds since epoch." }
        - { in: query, name: until, schema: { type: integer }, description: "Integer milliseconds since epoch." }
        - { in: query, name: limit, schema: { type: integer, minimum: 1 } }
        - { in: query, name: after_ms, schema: { type: integer }, description: "Keyset cursor; pair with after_id." }
        - { in: query, name: after_id, schema: { type: string }, description: "Keyset cursor; pair with after_ms." }
      responses:
        "200":
          description: A page of audit records, optionally with a `next` cursor.
          content:
            application/json:
              schema:
                type: object
                required: [records]
                properties:
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditRecord"
                  next:
                    type: object
                    required: [afterMs, afterId]
                    properties:
                      afterMs: { type: integer }
                      afterId: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/audit/{id}:
    get:
      tags: [audit]
      summary: One audit record by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: The matching audit record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sw_session
  responses:
    BadRequest:
      description: Request payload rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource does not exist (or is hidden from the caller's scope).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Credentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
    User:
      type: object
      required: [id, username, email, theme, createdAt, totpEnabled, disabled, mustChangePassword]
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        theme:
          type: string
          enum: ["", "light", "dark"]
        createdAt: { type: string, format: date-time }
        totpEnabled: { type: boolean }
        disabled: { type: boolean }
        disabledAt: { type: string, format: date-time }
        failedAttempts: { type: integer }
        lockedUntil: { type: string, format: date-time }
        mustChangePassword: { type: boolean }
...
    AuditRecord:
      type: object
      required: [id, occurredAt, actorKind, action, outcome]
      properties:
        id: { type: string }
        occurredAt: { type: string, format: date-time }
        actorKind:
          type: string
          enum: [user, system, unauthenticated]
        actorId: { type: string }
        actorLabel: { type: string }
        action: { type: string }
        targetKind: { type: string }
        targetId: { type: string }
        targetLabel: { type: string }
        outcome:
          type: string
          enum: [success, failure, denied]
        detail:
          type: object
          additionalProperties: true
        requestIp: { type: string }
        requestId: { type: string }